+++ Anarchy/REST/Message/Message/Components/Input/MessageInputComponent.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace Discord
+{
+    public class MessageInputComponent : MessageComponent
+    {
+        [JsonProperty("custom_id")]
+        public string Id { get; set; }
+
+        [JsonProperty("disabled")]
+        public bool Disabled { get; set; }
+    }
+}