+++ Anarchy/REST/Message/Message/Components/Input/SelectMenuOption.cs
@@ -1,0 +1,22 @@
+﻿using Newtonsoft.Json;
+
+namespace Discord
+{
+    public class SelectMenuOption
+    {
+        [JsonProperty("label")]
+        public string Text { get; set; }
+
+        [JsonProperty("value")]
+        public string Value { get; set; }
+
+        [JsonProperty("description")]
+        public string Description { get; set; }
+
+        [JsonProperty("emoji")]
+        public PartialEmoji Emoji { get; set; }
+
+        [JsonProperty("default")]
+        public bool? Default { get; set; }
+    }
+}