+++ Anarchy/WebSockets/Gateway/Guild/Member/List/MemberListItem.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+
+namespace Discord.Gateway
+{
+    public class MemberListItem : Controllable
+    {
+        public MemberListItem()
+        {
+            OnClientUpdated += (s, e) => Member.SetClient(Client);
+        }
+
+        [JsonProperty("member")]
+        public GuildMember Member { get; private set; }
+    }
+}