+++ Anarchy/WebSockets/Gateway/Guild/Member/Query/MemberListQueryOptions.cs
@@ -3,7 +3,6 @@
     public class MemberListQueryOptions
     {
         public int Offset { get; set; }
-
         public int Count { get; set; }
     }
 }