+++ Anarchy/WebSockets/Gateway/Interaction/Forms/ComponentFormInput.cs
@@ -1,0 +1,8 @@
+﻿namespace Discord.Gateway
+{
+    public abstract class ComponentFormInput
+    {
+        internal string Id { get; } = DiscordComponentForm.RandomString(8);
+        internal abstract void Handle(DiscordSocketClient client, DiscordInteraction interaction);
+    }
+}