+++ Anarchy/WebSockets/Gateway/Interaction/Forms/FormSelectMenuEventArgs.cs
@@ -1,0 +1,12 @@
+﻿namespace Discord.Gateway
+{
+    public class FormSelectMenuEventArgs : FormInteractionEventArgs
+    {
+        public string[] Values { get; }
+
+        public FormSelectMenuEventArgs(DiscordSocketClient client, DiscordInteraction interaction) : base(client, interaction)
+        {
+            Values = interaction.Data.SelectMenuValues;
+        }
+    }
+}